% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenDist.R
\name{DenDist}
\alias{DenDist}
\title{\eqn{L^2} Wasserstein distance between two distributions.}
\usage{
DenDist(d1 = NULL, d2 = NULL, q1 = NULL, q2 = NULL, qSup = NULL)
}
\arguments{
\item{d1, d2}{Two lists holding the density functions of the two distributions, each consisting of two numeric vectors \code{x} and \code{y} of the same length,
where \code{x} holds the support grid and \code{y} holds the values of the density function.
Note that \code{d1$x} and \code{d2$x} can have different lengths, and \code{d1} and \code{d2} overwrite \code{q1} and \code{q2}.}

\item{q1, q2}{Two numeric vectors holding the values of the quantile functions of the two distributions evaluated on the same grid given in \code{qSup}.}

\item{qSup}{A numeric vector holding the support grid of \code{q1} and \code{q2}; default: \code{seq(0,1,length(q1))}.}
}
\value{
A scalar holding the \eqn{L^2} Wasserstein distance between \code{d1} and \code{d2} or between \code{q1} and \code{q2}.
}
\description{
\eqn{L^2} Wasserstein distance between two distributions.
}
\examples{
d <- 3
y1 <- rnorm(d)
y1 <- y1 / sqrt(sum(y1^2))
y2 <- rnorm(d)
y2 <- y2 / sqrt(sum(y2^2))
dist <- SpheGeoDist(y1,y2)
}
