% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistCholesky.R
\name{DistCholesky}
\alias{DistCholesky}
\title{the Log-Cholesky and Cholesky distance between symmetric positive definite}
\usage{
DistCholesky(A = NULL, B = NULL, optns = list())
}
\arguments{
\item{A}{an p by p matrix}

\item{B}{an p by p matrix}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}. See `Details'.}
}
\value{
A list containing the following fields:
\item{dist}{the distance between matrices \code{A} and \code{B}.}
\item{optns}{A list containing the \code{optns} parameters utilized.}
}
\description{
the Log-Cholesky and Cholesky distance between two matrices
}
\details{
Available control options are 
\describe{
\item{metric}{Metric type choice, "log_cholesky", "cholesky" - default: \code{log_cholesky} for log Cholesky metric}
}
}
