% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expSphere.R
\name{expSphere}
\alias{expSphere}
\title{Compute an exponential map for a unit hypersphere.}
\usage{
expSphere(base, tg)
}
\arguments{
\item{base}{A unit vector of length \eqn{m} holding the base point of the tangent space.}

\item{tg}{A vector of length \eqn{m} of which the exponential map is taken.}
}
\value{
A unit vector of length \eqn{m}.
}
\description{
Compute an exponential map for a unit hypersphere.
}
