% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{frechet}
\alias{frechet}
\alias{frechet-package}
\title{frechet: Statistical Analysis for Random Objects and Non-Euclidean Data}
\description{
frechet: Statistical Analysis for Random Objects and Non-Euclidean Data
}
