% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFRCovCholesky.R
\name{LFRCovCholesky}
\alias{LFRCovCholesky}
\title{Local Fréchet regression of conditional covariance matrices with Log-Cholesky and Cholesky metric}
\usage{
LFRCovCholesky(x, M, xout, optns = list())
}
\arguments{
\item{x}{an n by p matrix of predictors.}

\item{M}{an q by q by n array (resp. a list of q by q matrices) where \code{M[,,i]} (resp. \code{M[[i]]}) contains the i-th covariance matrix of dimension q by q.}

\item{xout}{an m by p matrix of output predictor levels.}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}. See `Details'.}
}
\value{
A list containing the following fields:
\item{xout}{An m by p matrix of output predictor levels.}
\item{Mout}{A list of estimated conditional covariance matrices at \code{xout}.}
\item{opts}{A list containing the \code{opts} parameters utilized.}
}
\description{
Local Fréchet regression of covariance matrices with Euclidean predictors.
}
\details{
Available control options are 
\describe{
\item{bwCov}{bandwidth for conditional covariance estimation. If \code{bwCov} is not provided, it is chosen by cross validation.}
\item{kernel}{Name of the kernel function to be chosen from 'gauss', 'rect', 'epan', 'gausvar' and 'quar'. Default is 'gauss'.}
\item{corrOut}{Boolean indicating if Mout is shown as correlation or covariance matrix. Default: \code{FALSE} for only a covariance matrix.}
\item{metric}{Metric type choice, "log_cholesky", "cholesky" - default: \code{log_cholesky} for log Cholesky metric}
}
}
\examples{
n=30 #sample size
m=5 # dimension of covariance matrices
x=cbind(matrix(rnorm(n),n),matrix(rnorm(n),n)) #vector of predictor values
M <- array(0,c(m,m,n))
a = rnorm(m); b = rnorm(m)
A = diag(m)+a\%*\%t(a);
B = diag(m)+3*b\%*\%t(b);
for (i in 1:n){
  aux <- x[i,1]*A + x[i,2]**2*B
  M[,,i] <- aux \%*\% t(aux)
}
xout=cbind(runif(5),runif(5)) #output predictor levels
Covlist = LFRCovCholesky(x,M,xout)

}
\references{
\cite{A Petersen and HG Müller (2019). "Fréchet regression for random objects with Euclidean predictors." An. Stat. 47, 691-719.}
\cite{Z Lin (2019). " Riemannian Geometry of Symmetric Positive Definite Matrices via Cholesky Decomposition." Siam. J. Matrix. Anal, A. 40, 1353–1370.}
}
