% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logSphere.R
\name{logSphere}
\alias{logSphere}
\title{Compute a log map for a unit hypersphere.}
\usage{
logSphere(base, x)
}
\arguments{
\item{base}{A unit vector of length \eqn{m} holding the base point of the tangent space.}

\item{x}{A unit vector of length \eqn{m} which the log map is taken.}
}
\value{
A tangent vector of length \eqn{m}.
}
\description{
Compute a log map for a unit hypersphere.
}
