% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpheGeoGrad.R
\name{SpheGeoGrad}
\alias{SpheGeoGrad}
\title{Compute gradient w.r.t. y of the geodesic distance \eqn{\arccos(x^\top y)} on a unit hypersphere}
\usage{
SpheGeoGrad(x, y)
}
\arguments{
\item{x, y}{Two unit vectors.}
}
\value{
A vector holding gradient w.r.t. \code{y} of the geodesic distance between \code{x} and \code{y}.
}
\description{
Compute gradient w.r.t. y of the geodesic distance \eqn{\arccos(x^\top y)} on a unit hypersphere
}
