% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpheGeoHess.R
\name{SpheGeoHess}
\alias{SpheGeoHess}
\title{Hessian \eqn{\partial^2/\partial y \partial y^\top} of the geodesic distance \eqn{\arccos(x^\top y)} on a unit hypersphere}
\usage{
SpheGeoHess(x, y)
}
\arguments{
\item{x, y}{Two unit vectors.}
}
\value{
A Hessian matrix.
}
\description{
Hessian \eqn{\partial^2/\partial y \partial y^\top} of the geodesic distance \eqn{\arccos(x^\top y)} on a unit hypersphere
}
