% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WassFIntegral.R
\name{WassFIntegral}
\alias{WassFIntegral}
\title{Generalized Fréchet integrals of 1D distribution}
\usage{
WassFIntegral(phi, t_out, Q, Qout)
}
\arguments{
\item{phi}{An eigenfunction along which we want to project the distribution}

\item{t_out}{Support of \code{phi}}

\item{Q}{A \code{length(t_out)} X \code{length(Qout)} matrix whose jth row corresponds to the quantile function on grid \code{Qout} for the jth time point.}

\item{Qout}{Support of the quantile valued process}
}
\value{
A list of the following:
\item{f}{Quantile function corresponding to the frechet integral of \code{Q} along \code{phi}}
}
\description{
Calculating generalized Fréchet integrals of 1D distribution (equipped with Wasserstein distance)
}
\references{
\cite{Dubey, P., & Müller, H. G. (2020). Functional models for time‐varying random objects. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 82(2), 275-327.}
}
